requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@7ad00750c6422d8f19c0fb1859b7ee69fd4fa519
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
requests==2.28.0
urllib3==1.26.9
yarl==1.7.2
